backend/requirements.txt
Django==3.0.5
django-colorfield==0.4.5
djangorestframework==3.11.0
gunicorn==20.1.0
six==1.16.0
django-filter==2.4.0
pillow==8.4.0